+++ derive-encode/tests/lib.rs
@@ -211,4 +211,5 @@ fn build() {
     let t = trybuild::TestCases::new();
     t.pass("tests/build/redefine-prelude-symbols.rs");
     t.pass("tests/build/keep-impl-generics.rs");
+    t.compile_fail("tests/build/friendly-compilation-error-msg.rs");
 }